+++ src/runtime_src/driver/xclng/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_subdirectory(xrt)
 add_subdirectory(tools/xbflash)
-add_subdirectory(tools/xbsak)
+add_subdirectory(tools/xbutil)
