+++ src/runtime_src/driver/xclng/tools/xbutil/xbsak.sh
@@ -6,4 +6,4 @@ echo "!!Xbsak is obsolete and will be removed soon. Please use xbutil, instead!!
 echo ""
 
 PATH_DIR=`dirname $0`
-${PATH_DIR}/xbutil "$@"
+exec ${PATH_DIR}/xbutil "$@"
